// professional/src/app/models/person.ts
export interface Person {
  Photo: string;
  FirstName: string;
  SecondName: string;
  FirstLastName: string;
  SecondLastName: string;
  Age: number;
  Profession: string;
  Height: number;
  Weight: number;
  FootSize: number;
  Email: string;
  Address: string;
  PhoneNumber: string;
  Birth: Date;
  PlaceBirth: string;
  Nationality: string;
  Gender: string;
}